requirements.txt
asgiref==3.4.1
cycler==0.10.0
Django==3.2.9
django-crispy-forms==1.13.0
django-extensions==3.0.2
graphviz==0.14
kiwisolver==1.3.2
matplotlib==3.4.3
mysql-connector-python==8.0.20
mysqlclient==2.1.0
numpy==1.21.2
Pillow==8.3.2
protobuf==3.12.2
pydot==1.4.1
PyMySQL==1.0.2
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
tzdata==2021.5
